feat: Add support for specifying whisper model size
This commit adds support for specifying the size of the whisper model to use when transcribing audio files. The `--whispermodel` argument has been added to the command-line interface, allowing users to choose between different sizes of the model. By default, the "base" model is used if no size is specified.

The `transcribe()` function in `wspr_transcribe.py` has been updated to accept an optional `model_size` parameter. If a size is provided, it will be used to load the corresponding whisper model before transcribing audio files.

Note that downloading functionality has been commented out temporarily and will not be executed in this commit.