feat: Add support for LemonFox's whisper API
This commit adds support for using LemonFox's whisper API to transcribe audio files. The code now includes functions to split audio files into sizes of 100mb or lower, as per the API requirements. The transcribe function has been updated to call the new_transcribe function with the appropriate parameters based on whether the --whisperapi or --lemonfoxapi flag is used.

The main.py file has also been modified to include a new command line argument, --lemonfoxapi, which can be used to specify that the LemonFox's whisper API should be used for transcription.

Note: This commit does not include any changes related to specific names or files from the code.